fix(web): rename IssuesHint to IssuesAlert
