+++ web/src/components/software/SoftwarePage.tsx
@@ -34,7 +34,7 @@ import {
   Spinner,
   Stack,
 } from "@patternfly/react-core";
-import { Link, IssuesHint, Page } from "~/components/core";
+import { Link, Page, IssuesAlert } from "~/components/core";
 import UsedSize from "./UsedSize";
 import { useIssues } from "~/queries/issues";
 import {
@@ -163,7 +163,7 @@ function SoftwarePage(): React.ReactNode {
       <Page.Content>
         <Grid hasGutter>
           <GridItem sm={12}>
-            <IssuesHint issues={issues} />
+            <IssuesAlert issues={issues} />
           </GridItem>
           {showReposAlert && (
             <GridItem sm={12}>
